+++ backend/src/daoBean/RadnoMestoDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.RadnoMesto;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface RadnoMestoDaoLocal extends GenericDao<RadnoMesto, Long> {
+public interface RadnoMestoDaoLocal extends GenericDao<RadnoMesto, Integer> {
 
 }
