+++ backend/src/daoBean/SektorDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Sektor;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface SektorDaoLocal extends GenericDao<Sektor, Long>{
+public interface SektorDaoLocal extends GenericDao<Sektor, Integer>{
 
 }
