+++ backend/src/daoBean/ZaposleniDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.Zaposleni;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface ZaposleniDaoLocal extends GenericDao<Zaposleni, Long> {
+public interface ZaposleniDaoLocal extends GenericDao<Zaposleni, Integer> {
 
 }
